requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
cryptography==41.0.1
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
orjson==3.9.1
passlib==1.7.4
platformdirs==3.8.0
pre-commit==3.3.3
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
websockets==11.0.3
python-jose
